requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
boto==2.42.0
click==6.6
coreapi==1.32.0
decorator==4.0.10
dj-database-url==0.4.1
Django==1.9.8
gunicorn==19.6.0
psycopg2==2.6.2
dj-static==0.0.6
django-extensions==1.6.7
django-require==1.0.10
django-require-s3==1.0.0
django-storages==1.4.1
django-toolbelt==0.0.1
djangorestframework==3.4.0
Markdown==2.6.6
ordereddict==1.1
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
poster==0.8.1
pprintpp==0.2.3
prompt-toolkit==1.0.3
ptyprocess==0.5.1
pytz==2016.6.1
requests==2.10.0
sh==1.11
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
static3==0.7.0
traitlets==4.2.2
Unirest==1.1.7
uritemplate==0.6
wcwidth==0.1.7
whitenoise==3.2